feat: Use proper HTTP methods

+++ src/pages/api/products/[sku]/delete/index.ts
@@ -3,8 +3,10 @@ import { get } from "http";
 import { NextApiRequest, NextApiResponse } from "next"
 
 const handler = async (req: NextApiRequest, res: NextApiResponse) => {
-  await ProductModel.deleteProductBySku(req.query.sku as string);
-  res.status(200).json({ message: 'Product deleted' });
+  if (req.method === 'DELETE') {
+    await ProductModel.deleteProductBySku(req.query.sku as string);
+    res.status(200).json({ message: 'Product deleted' });  
+  }
 }
 
 export default handler;
